requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.2.2
pdfplumber==0.10.3
python-docx==1.1.0
pydantic==2.5.0
requests==2.31.0
numpy<2.0
openai>=1.0.0

